requirements.txt
torch
torchvision~=0.16.2

pillow~=10.0.1
matplotlib~=3.8.0
numpy~=1.26.3